I am stuck with models generation because Model Mommy has some limitations.
There is an open issue on the topic, and I will switch to the API development while the issue is being solve (I hope it does), in any case I will come back to this later.

The open issue is: https://github.com/vandersonmota/model_mommy/issues/144

And the summary of my problema and limitations is this:

Any progress on this issue?

I explain my case to let you know what I am missing the most in Model Mommy and I think is very related with this open issue. Also I apologize if the problems I have can be solved with Model Mommy as it is right now, I am quite a noob with Django and Python.

I have the following recipes:

user = Recipe(
    ChUser,
    date_joined=get_random_stuff.get_random_date,
    email=get_random_stuff.get_random_email,
    username=get_random_stuff.get_random_username,
)

tag = Recipe(
    TagModel,
    tag=get_random_stuff.get_random_tag,
)

category = Recipe(
    ChCategory,
    code=get_random_stuff.get_random_category_code,
)

hive = Recipe(
    ChHive,
    category=foreign_key(category),
    creation_date=get_random_stuff.get_random_date,
    creator=foreign_key(user),
    name=get_random_stuff.get_random_hive_name,
    ...
    tags=tags=related(tag, tag, tag),
)

I am using my own generators to fill the fields and I use make_recipe with the _quantity parameter to generate lots of objects with just one recipe definition (in my opinion the coolest feature of Model Mommy)

    The category objects are being created with a migration and they are a limited number of 25 categories. In the category recipe I just wanted get the code (randomly chosen between the 25 avaliable) of an existing category and automatically get the related existing category.
    I would like Mommy to just associate new hives with existent categories this way, but not to create new categories. Can I do that?

    In a similar manner, I'd like to always assign a creator to a new hive being this creator one of the existing users previously created using the user recipe.

    Last I would love to be able to associate some tags to the hive (being tags a many-to-many field) and this tags could be created be Mommy if they don't exist or just associated to the hive if they exist (the get_random_stuff.get_random_tag generator can give a new tag name or an existing one).
    Even more I would like to give a random number of tags associated to the hive (instead of always 3 like in the example) would that be possible?

Thanks a lot for your time and your effort in this project!